pyproject.toml
[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"

[installer]
no-binary = [ ":none:" ]

[tool.poetry]
name = "certilizer"
version = "1.1.1-pre.0"
description = "Generate report of SSL/TLS certificates from a list of endpoints defined in a YAML configuration file"
license = "Apache-2.0"
authors = [ "Cliffano Subagio <cliffano@gmail.com>" ]
readme = "README.md"
homepage = "https://github.com/cliffano/certilizer"
repository = "https://github.com/cliffano/certilizer"
documentation = "https://github.com/cliffano/certilizer"
keywords = [ "certilizer", "report", "ssl", "tls", "certificate" ]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent"
]

  [[tool.poetry.packages]]
  include = "certilizer"

  [tool.poetry.scripts]
  certilizer = "certilizer:cli"

  [tool.poetry.dependencies]
  python = "^3.10"
  cfgrw = "~1.0.0"
  click = "~8.2.1"
  conflog = "~2.0.1"
  dominate = "~2.9.1"
  pandas = "~2.3.0"
  PyYAML = "~6.0.2"
  tabulate = "~0.9.0"

[tool.poetry.group.dev.dependencies]
black = "~25.1.0"
twine = "~6.1.0"
wheel = "~0.45.1"

[tool.poetry.group.test.dependencies]
coverage = "~7.8.2"
numpy = "1.26.4"
pylint = "~3.3.7"
pytest = "~8.3.5"
wily = "~1.25.0"

[tool.poetry.group.doc.dependencies]
pylint-report = "~2.4.2"
pytest-html = "~4.1.1"
Sphinx = "8.1.3"
